Move `DataHandler` and `InvocationProcess` classes to their own module, Rename `board_request` to `allocate_cache` since that's all it does
